Move account based indexing methods to separate file
